+++ backend/apps/users/api/views.py
@@ -1,5 +1,5 @@
 from django.contrib.auth import authenticate, get_user_model
-from rest_framework import status
+from rest_framework import permissions, status, views
 from rest_framework.generics import get_object_or_404
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -10,10 +10,15 @@ User = get_user_model()
 
 
 class CustomTokenObtainPairView(TokenObtainPairView):
+
     def post(self, request: Request, *args, **kwargs) -> Response:
         serializer = self.get_serializer(data=request.data)
-        email_ = request.data["email"]
-        password = request.data["password"]
+
+        try:
+            email_ = request.data["email"]
+            password = request.data["password"]
+        except KeyError:
+            return Response({"msg": "email and/or password not provided."}, status=status.HTTP_400_BAD_REQUEST)
 
         if authenticate(email=email_, password=password) is None:
             user = get_object_or_404(User, email=email_)
@@ -21,14 +26,7 @@ class CustomTokenObtainPairView(TokenObtainPairView):
             if not user.is_active:
                 return Response({"msg": "user is not active."}, status=status.HTTP_401_UNAUTHORIZED)
 
-            try:
-                User.objects.get(email=email_, is_active=True)
-            except User.DoesNotExist:
-                return Response(
-                    {"msg": "user with this email does not exist."},
-                    status=status.HTTP_404_NOT_FOUND,
-                )
-            return Response({"msg": "user password wrong."}, status=status.HTTP_403_FORBIDDEN)
+            return Response({"msg": "user password wrong."}, status=status.HTTP_401_UNAUTHORIZED)
 
         try:
             serializer.is_valid(raise_exception=True)
@@ -41,3 +39,31 @@ class CustomTokenObtainPairView(TokenObtainPairView):
         response.set_cookie("access_token", serializer.validated_data["access"])
         return response
         # return Response(serializer.validated_data, status=status.HTTP_200_OK)
+
+
+class SpamEmailEveryWeek(views.APIView):
+    """
+    Send spam emails every week.
+    This class allows users to subscribe and unsubscribe from a weekly newsletter.
+    """
+
+    permission_classes = [permissions.IsAuthenticated]
+    serializer_class = None
+
+    def post(self, request):
+        user = request.user
+        if not user.is_spam_email:
+            user.is_spam_email = True
+            user.save()
+            return Response({"msg": "You subscribed to the newsletter"}, status.HTTP_200_OK)
+
+        return Response({"msg": "You are already subscribed to the newsletter"}, status.HTTP_200_OK)
+
+    def delete(self, request):
+        user = request.user
+        if user.is_spam_email:
+            user.is_spam_email = False
+            user.save()
+            return Response({"msg": "You unsubscribed from the newsletter"}, status.HTTP_200_OK)
+
+        return Response({"msg": "You are not subscribed to the newsletter"}, status.HTTP_200_OK)